Please Check my Jade: Added invisible td
Something tells me that the {} brackets are unnecessary...but I want to be clear that the if statement doesn't extend to the following lines.
Can someone who understands jade please fix this and/or tell me if it breaks anything? Thanks :)

It should: insert <td style="display:none" class=""></td> before the third H block td.