Hopefully last commit related to H block bugs
I came up with a few ways to fix it, and this way seemed the least likely to break anything. Let me know if it does.

The problem was that an event saved on Thursday showed up on Friday when refreshed. It was caused by a discrepancy between the week structure, and the saved events variable. (events took into account the lack of 6th period on wednesday, the week structure did not). One way to fix it would be to change the week structure, and then handle the blanks in the jade file. This would probably break stuff so I decided to fix it in as small a location as I could (changing only a variable local to a loop).